split up find
start to work on findHtml
