+++ packages/browser/tests/actions/click.test.ts
@@ -1,23 +1,19 @@
 import { CONFIG } from "@qawolf/config";
 import { Browser } from "../../src/Browser";
 import { click } from "../../src/actions";
-import { hasText } from "../../src/find";
+import { hasText } from "../../src/find/hasText";
 
 describe("click", () => {
   it("clicks on link", async () => {
     const browser = await Browser.create({ url: CONFIG.testUrl });
     const page = await browser.currentPage();
 
-    const element = await browser.element({
-      action: "click",
-      index: 0,
-      target: {
-        innerText: "broken images",
-        xpath: '//*[@id="content"]/ul/li[3]/a'
-      }
-    });
-
+    // TODO targetString -> (html/css)
+    // TODO click(targetString)
+    // TODO Browser.element step -> locator (html?, css?, action?, page?)
+    const element = await browser.find({ text: "broken images" });
     await click(element);
+
     await page.waitForNavigation();
     expect(page.url()).toBe(`${CONFIG.testUrl}broken_images`);
 
@@ -35,15 +31,7 @@ describe("click", () => {
     );
     expect(hasInvalidUsernameText).toBe(false);
 
-    const element = await browser.element({
-      action: "click",
-      index: 0,
-      target: {
-        innerText: "login",
-        tagName: "i",
-        xpath: "//*[@id='login']/button/i"
-      }
-    });
+    const element = await browser.find({ html: "<i>Login</i>" });
     await click(element);
 
     const hasInvalidUsernameText2 = await hasText(page, "username is invalid");
